requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
argparse==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
httplib2==0.31.2
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
pip==25.3
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyparsing==3.3.2
python-dotenv==1.2.2
redis==7.3.0
requests==2.32.5
rich==14.3.3
shellingham==1.5.4
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uritemplate==4.2.0
urllib3==2.6.3
uv==0.10.12
